@@ -1142,10 +1142,10 @@ class GrockGlassMorphism extends StatelessWidget {
         filter: ImageFilter.blur(sigmaX: blur, sigmaY: blur),
         child: Container(
           decoration: BoxDecoration(
-            color: color.withOpacity(opacity),
+            color: color.withValues(alpha:opacity),
             borderRadius: BorderRadius.circular(borderRadius ?? 10),
             border: Border.all(
-              color: color.withOpacity(0.1),
+              color: color.withValues(alpha:0.1),
               width: 0.5,
             ),
           ),
