@@ -158,7 +158,7 @@ class _GrockToastWidgetState extends State<_GrockToastWidget>
                         boxShadow: widget.boxShadow ??
                             [
                               BoxShadow(
-                                color: Colors.black.withOpacity(0.2),
+                                color: Colors.black.withValues(alpha:0.2),
                                 blurRadius: 15,
                               ),
                             ],
