@@ -1,9 +1,15 @@
 enum NavType {
+  material,
+  cupertino,
   fade,
-  leftToRight,
-  rightToLeft,
-  topToBottom,
-  bottomToTop,
-  rightToLeftWithFade,
-  leftToRightWithFade,
+  slideLeft,
+  slideRight,
+  slideUp,
+  slideDown,
+  scale,
+  rotate,
+  size,
+  fadeScale,
+  flip,
+  ripple,
 }
